requirements.txt
requests~=2.31.0
bs4~=0.0.1
beautifulsoup4~=4.11.1
discord~=2.0.0
python-dotenv~=0.20.0
pyttsx3~=2.90
pedalboard~=0.8.7
mutagen~=1.45.1
setuptools~=60.2.0